load color asyncly
